Ensure 'package' directory exists before deleting empty dirs
Previously, the script attempted to delete empty directories without checking if the 'package' directory existed, potentially leading to errors. This update adds a conditional check to verify the existence of the 'package' directory before running the delete command, improving script robustness.

+++ scripts/make-packages.sh
@@ -84,7 +84,9 @@ function make_package() {
     fi
 
     # remove empty directories e.g. examples if this is left empty after the previous move
-    find package -type d -empty -delete 2>&1 >> /dev/null
+    if [[ -d package ]]; then
+      find package -type d -empty -delete 2>&1 >> /dev/null
+    fi
 
     # copy the IG resource here because bake refuses to do that
     # cp ./ig/*.json ./package/
